refactor(1353): extract view models and widgets
